Add basic Podman support
Previously, with Podman installed in rootless mode, it-depends
crashes telling me I need to `apt-get install` docker.

Add some basic conditional logic to check for a Podman rootless socket
and use it if present. If not found, _try_podman_socket() returns
`None`, the previous value supplied by it-depends for base_url. This
should make it backwards-compatible with existing it-depends users.